requirements.txt
flask
azure-storage-blob
azure-ai-formrecognizer
python-dotenv
gunicorn
azure-data-tables